Backend/requirements.txt
fastapi==0.110.0
uvicorn==0.28.0
openai==1.50.0
python-dotenv==1.0.1
pypdf==4.1.0
python-multipart==0.0.9
httpx==0.27.0